refactoring search on the backend/python side.  Searches return "hits" now rather than seperate arrays of data objects.  Hits contain title, content, date, and type information, along with highlights to displaying where a term occurs in the content if applicable.
We now index "documents" rather than "resources", which was an overloaded term.  Study, Training, and Resource models now all provide content that is indexed in ElasticSearch as a Document object, which is returned and converted to a Hit with highlighting when returning via the API.

Upgraded ElasticSearch libraries to 6.3.1.
Fixed the tests.