The age ranges and categories applied in the CSV / Spreadsheets were inconsistent.  For the age ranges I forced them to be consistent, editing the online spreadsheet so they are defined the samy way for studies, locations, resources and events.   For categories, I modified the data loader so that it will fail if a category shows up on a resource or study that was not defined in the category csv.
Age Ranges are no longer categories.  They are loaded into the database and elastic search in their own field "ages" (thus the migration)

Facets are now properly called "aggregations" and show up in the search as a deeply nested dictionary. They only apply to ages and types presently.  Everthing else is handled as a topic.

Ages and Types are now top level arrays on the Search model, and can be used to filter searches.  There is not generic "filter" anymore.  These are what you get.

Upgraded marshmallow to version 3 to get some clearer bug free serialization of dictionaries.